A3_part1

+++ MockitoProject/src/main/java/IA3Part1.java
@@ -1,6 +1,6 @@
 
 public interface IA3Part1 {
-	public double computeTax (double amount);
+	public double computeTax (int amount);
 	public double weeklySalary (Employee income);
 	public double Salaryfortnight (Employee income);
 	public double computeKiwiSaver (Employee KiwiOption);
